Use the Profile service/store for the enrollment profile.
This allows the custom modification of the enrollment profile by the end user. However as a trade off we now essentially store a "static" enrollment profile. This is intended to change in the future when "templated" profiles are implemented where dynamic substring replacement (for example) is allowed in a Mobileconfig. This will allow for more dyanmic-yet-custom profiles in the case of e.g. per-device SCEP challenges.