add deptoken to the platform folder.
I noticed that all deptoken is is a store for DEP tokens, which fits nicely with
the rest of the storage/api packages in platform.

The next step will be to move the deptoken store into the config package, but it must be
done in a gradual way since people are already using it today.