Use id set as parameter in checkin URL and send JSON webhook callback
If an id parameter is specified in the enroll URL /mdm/enroll?id=123,
then use this in the checkin url /mdm/checkin?id=123 in the mdm profile.

If checkin webhook is specified, the id and device id (UDID) will be
encoded as JSON and sent when the device check in and out.

The JSON webhook is generic and can be used for other purposes than
checkin events.